[TASK] Move missing permissions calculation into separate file
